chore: setup Typescript
