+++ .eslintignore
@@ -2,3 +2,4 @@ packages/java-parser/samples
 packages/prettier-plugin-java/samples
 packages/prettier-plugin-java/test-samples
 node_modules
+packages/prettier-plugin-java/dist
