+++ package.json
@@ -15,7 +15,8 @@
     "lint": "eslint packages/**/*.js",
     "format:fix": "prettier --write \"**/*.@(js|json|md)\"",
     "format:validate": "prettier --list-different \"**/*.@(js|json|md)\"",
-    "build": "node packages/java-parser/scripts/unicode.js packages/java-parser/resources/Unicode/UnicodeData.txt && prettier --write packages/java-parser/src/unicodesets.js && node packages/java-parser/scripts/gen-diagrams.js",
+    "build": "yarn build:prettier-plugin-java && node packages/java-parser/scripts/unicode.js packages/java-parser/resources/Unicode/UnicodeData.txt && prettier --write packages/java-parser/src/unicodesets.js && node packages/java-parser/scripts/gen-diagrams.js",
+    "build:prettier-plugin-java": "cd packages/prettier-plugin-java && yarn build && cd ../..",
     "prettier": "prettier --plugin=./packages/prettier-plugin-java --parser=java",
     "update-test-outputs": "node packages/prettier-plugin-java/scripts/update-test-output.js"
   },
