+++ packages/prettier-plugin-java/package.json
@@ -2,7 +2,7 @@
   "name": "prettier-plugin-java",
   "version": "1.3.1",
   "description": "Prettier Java Plugin",
-  "main": "src/index.js",
+  "main": "dist/cjs/index.js",
   "repository": "https://github.com/jhipster/prettier-java",
   "license": "Apache-2.0",
   "dependencies": {
@@ -12,12 +12,19 @@
   },
   "scripts": {
     "test": "npm-run-all test:unit test:e2e-core",
-    "test:unit": "mocha \"test/unit-test/**/*.spec.js\" \"test/unit-test/**/*-spec.js\"",
-    "test:unit:coverage": "nyc mocha \"test/unit-test/**/*.spec.js\" \"test/unit-test/**/*-spec.js\"",
-    "test:e2e-core": "node scripts/clone-samples e2e-core && mocha \"test/repository-test/core-test.js\"",
-    "test:e2e-jhipster1": "node scripts/clone-samples e2e-jhipster1 && mocha \"test/repository-test/jhipster-1-test.js\"",
-    "test:e2e-jhipster2": "node scripts/clone-samples e2e-jhipster2 && mocha \"test/repository-test/jhipster-2-test.js\"",
+    "test:unit": "yarn build && mocha \"test/unit-test/**/*.spec.js\" \"test/unit-test/**/*-spec.js\"",
+    "test:unit:coverage": "yarn build && nyc mocha \"test/unit-test/**/*.spec.js\" \"test/unit-test/**/*-spec.js\"",
+    "test:e2e-core": "yarn build && node scripts/clone-samples e2e-core && mocha \"test/repository-test/core-test.js\"",
+    "test:e2e-jhipster1": "yarn build && node scripts/clone-samples e2e-jhipster1 && mocha \"test/repository-test/jhipster-1-test.js\"",
+    "test:e2e-jhipster2": "yarn build && node scripts/clone-samples e2e-jhipster2 && mocha \"test/repository-test/jhipster-2-test.js\"",
     "test:all": "npm-run-all test test:e2e-jhipster1 test:e2e-jhipster2",
-    "clone-samples": "node scripts/clone-samples.js"
+    "clone-samples": "node scripts/clone-samples.js",
+    "build": "tsc -p tsconfig.json --outDir dist/cjs"
+  },
+  "devDependencies": {
+    "@chevrotain/types": "9.0.2",
+    "@types/lodash": "4.14.172",
+    "@types/prettier": "2.3.1",
+    "typescript": "^4.3.5"
   }
 }
