+++ packages/prettier-plugin-java/test/unit-test/snippets/types-values-and-variables/wildcard/test.spec.js
@@ -2,7 +2,9 @@
 
 const { assert, spy } = require("sinon");
 const { expectSnippetToBeFormatted } = require("../../../../test-utils");
-const { CstPrettierPrinter } = require("../../../../../src/cst-printer");
+const {
+  BaseCstPrettierPrinter
+} = require("../../../../../dist/cjs/base-cst-printer");
 
 describe("Wildcard", () => {
   let wildcardBoundsSpy;
@@ -12,7 +14,7 @@ describe("Wildcard", () => {
     if (wildcardBoundsSpy !== undefined) {
       wildcardBoundsSpy.restore();
     }
-    wildcardBoundsSpy = spy(CstPrettierPrinter.prototype, "wildcardBounds");
+    wildcardBoundsSpy = spy(BaseCstPrettierPrinter.prototype, "wildcardBounds");
   });
 
   it("can format a wildcard", () => {
